+++ rpc-responses/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "solders-rpc-responses"
+version = "0.14.4"
+edition = "2021"
+include = ["/src"]
+license = "MIT"
+repository = "https://github.com/kevinheavey/solders"
+
+[dependencies]
+pyo3 = { workspace = true }
+serde = { workspace = true }
+serde_with = { workspace = true }
+serde_json = { workspace = true }
+derive_more = { workspace = true }
+solders-macros = { workspace = true }
+solders-primitives = { workspace = true }
+solders-traits = { workspace = true }
+solders-account = { workspace = true }
+solders-account-decoder = { workspace = true }
+solders-transaction-status = { workspace = true }
+solders-rpc-common = { workspace = true }
+solders-rpc-errors = { workspace = true }
+solders-rpc-requests = { workspace = true }
+solana-sdk = { workspace = true }
+thiserror = { workspace = true }
+camelpaste = { workspace = true }
