+++ src/rpc/mod.rs
@@ -1,14 +1,11 @@
 use std::collections::HashMap;
 
-use self::responses::create_responses_mod;
 use pyo3::prelude::*;
 use solders_rpc_config::create_config_mod;
 use solders_rpc_errors::create_errors_mod;
 use solders_rpc_filter::create_filter_mod;
 use solders_rpc_requests::create_requests_mod;
-
-pub mod responses;
-mod tmp_response;
+use solders_rpc_responses::create_responses_mod;
 
 pub fn create_rpc_mod(py: Python<'_>) -> PyResult<&PyModule> {
     let rpc_mod = PyModule::new(py, "rpc")?;
